+++ docker-compose.yml
@@ -1,4 +1,4 @@
-version: '3.7'
+version: '3.3'
 services:
   nlp:
     image: docker.lappsgrid.org/askme/nlp
