+++ src/test/groovy/org/lappsgrid/askme/nlp/IntegrationTest.groovy
@@ -1,4 +1,4 @@
-package org.lappsgrid.eager.mining.web.nlp.stanford
+package org.lappsgrid.askme.nlp
 
 import org.junit.*
 import org.lappsgrid.askme.nlp.Main
