+++ src/test/groovy/org/lappsgrid/askme/nlp/PipelineTest.groovy
@@ -1,4 +1,4 @@
-package org.lappsgrid.eager.mining.web.nlp.stanford
+package org.lappsgrid.askme.nlp
 
 
 import org.junit.Test
