+++ src/test/groovy/org/lappsgrid/askme/nlp/StressTest.groovy
@@ -1,7 +1,6 @@
-package org.lappsgrid.eager.mining.web.nlp.stanford
+package org.lappsgrid.askme.nlp
 
 import groovy.util.logging.Slf4j
-import org.lappsgrid.askme.nlp.Main
 
 //import org.lappsgrid.eager.mining.core.Configuration
 //import org.lappsgrid.eager.mining.core.jmx.Registry
